refactor: split member dashboard view
