+++ apps/web/src/components/dashboard/member-dashboard-view/data.ts
@@ -1,0 +1,15 @@
+import { and, db, eq, subscriptions, user } from '@interdomestik/database';
+import { cache } from 'react';
+
+export const getCachedUser = cache(async (userId: string) => {
+  return db.query.user.findFirst({
+    where: eq(user.id, userId),
+  });
+});
+
+export const getCachedTenantSubscriptions = cache(async (userId: string, tenantId: string) => {
+  return db.query.subscriptions.findMany({
+    where: and(eq(subscriptions.userId, userId), eq(subscriptions.tenantId, tenantId)),
+    orderBy: (subscriptionTable, { desc }) => [desc(subscriptionTable.createdAt)],
+  });
+});
